Move binding diagnostics out of core
Drop the Ray\Bindings namespace (Bindings, BindingsMarkdown, BindingsHtml,
the bindings-html CLI and its viewer assets) and the deprecated
Ray\Di\BindingsMarkdown shim. Rich visualization moves to the
ray/object-visual-grapher package; composition provenance (BindingLog)
stays in core.

Ray\Di\ModuleVisitorInterface replaces Ray\Bindings\ModuleVisitorInterface
as the stable route for tools to visit a module's composed container via
AbstractModule::accept().
