--- bin/bindings-html
@@ -1,66 +1,0 @@
-#!/usr/bin/env php
-<?php
-
-declare(strict_types=1);
-
-/**
- * bindings-html — render a Ray.Di bindings.md as an HTML view.
- *
- * Usage:
- *   bin/bindings-html <path/to/bindings.md> [composer.lock] [message] > bindings.html
- *   bin/bindings-html - [composer.lock] [message] < bindings.md       > bindings.html
- *
- * The markdown is embedded verbatim in a <pre> data island; a shared viewer
- * (stylesheet + script served from jsDelivr) decorates it on load: colour-coded
- * events, a type filter, and the discarded side of every collision in red.
- * Without JavaScript the <pre> shows the plain log.
- *
- * Pass a composer.lock to link each class to its source — a GitHub URL for
- * humans and the local vendor/ path (data-src) for agents. The optional
- * [message] renders as a subtitle, e.g. the context the bindings were composed
- * in.
- *
- * The rendering lives in Ray\Bindings\BindingsHtml so every consumer — this
- * CLI, a documentation generator — reuses the same viewer.
- */
-
-use Ray\Bindings\BindingsHtml;
-
-$autoload = null;
-foreach ([__DIR__ . '/../vendor/autoload.php', __DIR__ . '/../../../autoload.php'] as $file) {
-    if (file_exists($file)) {
-        $autoload = $file;
-        break;
-    }
-}
-
-if ($autoload === null) {
-    fwrite(STDERR, "bindings-html: Composer autoloader not found; run `composer install`\n");
-    exit(1);
-}
-
-require $autoload;
-
-$path = $argv[1] ?? null;
-if ($path === null || $path === '-h' || $path === '--help') {
-    fwrite(STDERR, "usage: bindings-html <bindings.md|-> [composer.lock] [message] > out.html\n");
-    exit($path === null ? 1 : 0);
-}
-
-$md = $path === '-' ? file_get_contents('php://stdin') : @file_get_contents($path);
-if ($md === false) {
-    fwrite(STDERR, "bindings-html: cannot read {$path}\n");
-    exit(1);
-}
-
-$lock = '';
-$lockPath = $argv[2] ?? '';
-if ($lockPath !== '') {
-    $lock = @file_get_contents($lockPath);
-    if ($lock === false) {
-        fwrite(STDERR, "bindings-html: cannot read {$lockPath}\n");
-        exit(1);
-    }
-}
-
-echo (new BindingsHtml())->page($md, $lock, $argv[3] ?? '');
