+++ composer.json
@@ -9,7 +9,6 @@
             "email": "akihito.koriyama@gmail.com"
         }
     ],
-    "bin": ["bin/bindings-html"],
     "require": {
         "php": "^8.2",
         "koriym/null-object": "^1.0",
@@ -21,7 +20,8 @@
         "phpunit/phpunit": "^11.0"
     },
     "suggest": {
-        "ray/compiler": "For compiling dependency injection container to improve performance"
+        "ray/compiler": "For compiling dependency injection container to improve performance",
+        "ray/object-visual-grapher": "For visualizing composed bindings and the object graph"
     },
     "config": {
         "sort-packages": true,
@@ -32,14 +32,12 @@
     },
     "autoload": {
         "psr-4": {
-            "Ray\\Di\\": ["src/di", "src-deprecated/di"],
-            "Ray\\Bindings\\": "src-bindings/"
+            "Ray\\Di\\": ["src/di", "src-deprecated/di"]
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Ray\\Di\\": ["tests/di", "tests/di/Fake/"],
-            "Ray\\Bindings\\": "tests-bindings/"
+            "Ray\\Di\\": ["tests/di", "tests/di/Fake/"]
         }
     },
     "scripts": {
@@ -51,8 +49,8 @@
             "php -dextension=pcov.so -r \"extension_loaded('pcov') || (fwrite(STDERR, 'pcov required for mutation testing (dev-only). Install: pecl install pcov'.PHP_EOL) && exit(1));\"",
             "php -dextension=pcov.so -d pcov.enabled=1 vendor/bin/infection --threads=2 --min-msi=90 --min-covered-msi=90 --no-progress"
         ],
-        "cs": ["vendor-bin/tools/vendor/squizlabs/php_codesniffer/bin/phpcs --standard=./phpcs.xml src tests src-bindings tests-bindings"],
-        "cs-fix": ["vendor-bin/tools/vendor/squizlabs/php_codesniffer/bin/phpcbf src tests src-bindings tests-bindings"],
+        "cs": ["vendor-bin/tools/vendor/squizlabs/php_codesniffer/bin/phpcs --standard=./phpcs.xml src tests"],
+        "cs-fix": ["vendor-bin/tools/vendor/squizlabs/php_codesniffer/bin/phpcbf src tests"],
         "clean": ["vendor-bin/tools/vendor/phpstan/phpstan/phpstan clear-result-cache", "vendor-bin/tools/vendor/vimeo/psalm/psalm --clear-cache", "rm -rf tests/tmp/*.php"],
         "sa": ["vendor-bin/tools/vendor/vimeo/psalm/psalm -m -c psalm.xml --show-info=false", "vendor-bin/tools/vendor/phpstan/phpstan/phpstan analyse -c phpstan.neon --no-progress "],
         "metrics": ["@test", "vendor-bin/tools/vendor/phpmetrics/phpmetrics/bin/phpmetrics --report-html=build/metrics --exclude=Exception --log-junit=build/junit.xml --junit=build/junit.xml src"],
