+++ phpstan.neon
@@ -3,9 +3,7 @@ parameters:
   level: max
   paths:
     - src/di
-    - src-bindings
     - tests/di
-    - tests-bindings
     - tests/type
   excludePaths:
     - tests/tmp/*
@@ -17,5 +15,4 @@ parameters:
   ignoreErrors:
     - '#contains generic class ReflectionAttribute but does not specify its types:#'
     - '#generic class ReflectionAttribute but does not specify its types:#'
-    - '#Offset 1 might not exist on array\{\}\|array\{non-falsy-string, string\}#'
     - '#will always evaluate to true#'
