+++ phpunit.xml.dist
@@ -8,13 +8,11 @@
   <source>
     <include>
       <directory suffix=".php">src/di</directory>
-      <directory suffix=".php">src-bindings</directory>
     </include>
   </source>
   <testsuites>
     <testsuite name="Ray.Di Test Suite">
       <directory>tests/di/</directory>
-      <directory>tests-bindings</directory>
       <directory phpVersion="8.0.0" phpVersionOperator=">=">tests-php8</directory>
     </testsuite>
   </testsuites>
