+++ src/di/AbstractModule.php
@@ -8,7 +8,6 @@ use Ray\Aop\AbstractMatcher;
 use Ray\Aop\Matcher;
 use Ray\Aop\Pointcut;
 use Ray\Aop\PriorityPointcut;
-use Ray\Bindings\ModuleVisitorInterface;
 use Stringable;
 
 use function assert;
